Move subscriptions-checking to instrumenter
There are some call sites that are not using the high-level `Notifications`
module, but instead are calling the instrumenter directly
(e.g. `ActiveSupport::Notifications.instrumenter.instrument`).

This is an optimization for those call sites, so that they can avoid
the overhead of building event objects and collecting instrumentation
information if there are no subscribers.

Co-authored-by: Schlubbi Jöst <schlubbi@github.com>